+++ app/static/parts-ui/assets/ThreeMFViewer-Cd-abdMm.js
@@ -1,4 +1,4 @@
-import{r as Ue,j as et}from"./index-CqT-YUgg.js";/**
+import{r as Ue,j as et}from"./index-CKl3UoVM.js";/**
  * @license
  * Copyright 2010-2026 Three.js Authors
  * SPDX-License-Identifier: MIT